improve code
- fixed random enemy bug
- use scriptableObject to store enemy's data
- fixed HpBarUI to updateUI when player's HP change instead update every time
- fixed enemy groundChecker to flexible with enemy size instead use fix number

+++ Assets/Scripts/Menu/Save.cs
@@ -7,7 +7,7 @@ public class Save
 {
     public float playerPosX;
     public float playerPosY;
-    public int playerHp;
+    public float playerHp;
     
     public int currentBullet_0 = 0;
     public int currentBullet_1 = 0;
